Move all heatmap components into new dir heatmap/ for organization
