+++ src/components/profile/progress/heatmap/HeatMap.tsx
@@ -1,6 +1,6 @@
 import { Card, Stack, Typography } from "@mui/joy";
 import HeatMapTile from "./HeatMapTile";
-import HeatMapLegendTile from "./HeatMapLegendTile";
+import HeatMapLegend from "./HeatMapLegend";
 
 export default function HeatMap({ activity }: { activity: string[] }) {
   // make it total problems touched, ie unique problems submitted for
@@ -100,29 +100,7 @@ export default function HeatMap({ activity }: { activity: string[] }) {
             </Stack>
           ))}
         </Stack>
-        <Stack direction="row-reverse" alignItems="center" gap={0.25}>
-          <HeatMapLegendTile
-            title="75-100% of most active day"
-            heatmapColour={colours[4]}
-          />
-          <HeatMapLegendTile
-            title="50-74% of most active day"
-            heatmapColour={colours[3]}
-          />
-          <HeatMapLegendTile
-            title="25-49% of most active day"
-            heatmapColour={colours[2]}
-          />
-          <HeatMapLegendTile
-            title="0-24% of most active day"
-            heatmapColour={colours[1]}
-          />
-          <HeatMapLegendTile
-            title="No contributions"
-            heatmapColour={colours[0]}
-          />
-          <Typography level="body-sm" sx={{ marginRight: "5px" }}>Legend:</Typography>
-        </Stack>
+        <HeatMapLegend colours={colours}/>
       </Card>
     </Stack>
   )
